+++ lib/ethers/transaction.ex
@@ -19,7 +19,7 @@ defmodule Ethers.Transaction do
     gas_price: nil,
     hash: nil,
     max_fee_per_gas: nil,
-    max_priority_fee_per_gas: "0x0",
+    max_priority_fee_per_gas: nil,
     nonce: nil,
     signature_r: nil,
     signature_s: nil,
@@ -58,7 +58,7 @@ defmodule Ethers.Transaction do
   @common_fillable_params [:chain_id, :nonce]
   @type_fillable_params %{
     legacy: [:gas_price],
-    eip1559: [:max_fee_per_gas]
+    eip1559: [:max_fee_per_gas, :max_priority_fee_per_gas]
   }
   @integer_type_values [
     :block_number,
@@ -240,8 +240,9 @@ defmodule Ethers.Transaction do
   end
 
   defp fill_action(:chain_id, _tx), do: :chain_id
-  defp fill_action(:nonce, tx), do: {:get_transaction_count, [tx.from, "latest"]}
+  defp fill_action(:nonce, tx), do: {:get_transaction_count, tx.from, block: "latest"}
   defp fill_action(:max_fee_per_gas, _tx), do: :gas_price
+  defp fill_action(:max_priority_fee_per_gas, _tx), do: :max_priority_fee_per_gas
   defp fill_action(:gas_price, _tx), do: :gas_price
 
   defp post_process([], [], acc), do: {:ok, Enum.into(acc, %{})}
@@ -260,6 +261,15 @@ defmodule Ethers.Transaction do
       mex_fee_per_gas = div(v * 120, 100)
       {:ok, {:max_fee_per_gas, Utils.integer_to_hex(mex_fee_per_gas)}}
     end
+  end
+
+  defp do_post_process(:max_priority_fee_per_gas, {:ok, v_int}) do
+    # use latest max_priority_fee_per_gas from the chain as default
+    {:ok, {:max_priority_fee_per_gas, Utils.integer_to_hex(v_int)}}
+  end
+
+  defp do_post_process(:nonce, {:ok, nonce}) when is_integer(nonce) do
+    {:ok, {:nonce, Utils.integer_to_hex(nonce)}}
   end
 
   defp do_post_process(key, {:ok, v_hex}) do